pyproject.toml
[project]
name = "weather-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "gradio-client>=2.0.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "litellm>=1.80.11",
    "peewee>=3.18.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
    "wikipedia>=1.4.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "jupyter>=1.1.1",
    "marimo>=0.18.4",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings-python>=2.0.1",
    "wat>=0.7.0",
]
